Changing type and adding checkAlertExists to analyzeAgentOccupancy

+++ src/main/java/tc3005b224/amazonconnectinsights/repository/AlertRepository.java
@@ -30,5 +30,5 @@ public interface AlertRepository extends JpaRepository<Alert, Long> {
 
     Iterable<Alert> findByConnectionIdentifierAndResourceAndSolvedAndHasTraining(int connectionIdentifier, String resource, Boolean solved, Boolean hasTraining);
 
-    Iterable<Alert> findByResourceAndInsightIdentifierAndDateRegisteredBetween(String resource, Long insightIdentifier, Date dateFrom, Date dateTo);
+    Iterable<Alert> findByResourceAndInsightIdentifierAndDateRegisteredBetween(String resource, Short insightIdentifier, Date dateFrom, Date dateTo);
 }
